Feb 24: Major refactoring and documentation updates
- Refactor: Split scripts into train/ and inference/ folders
- Refactor: Remove v6_ prefix from all script names
- Refactor: Move train/ and inference/ to repo root
- Feat: Add MATH 7B/72B hybrid pipeline scripts
- Feat: Add MATH dataset scripts and batch evaluation tools
- Feat: Refactor e2e_pipeline with bridging search
- Feat: Add Track 1 algebraic solver with sympy.solve()
- Docs: Update README with lambda epigraph, error attribution, lessons learned
- Docs: Add field guide, training pipeline, and paper docs
- Docs: Simplify plan folder file names

+++ train/train_classifier.py
@@ -100,7 +100,7 @@ class WeightedTrainer(Trainer):
         outputs = model(**inputs)
         logits = outputs.logits
 
-        loss_fct = nn.CrossEntropyLoss(weight=self.class_weights.to(logits.device))
+        loss_fct = nn.CrossEntropyLoss(weight=self.class_weights.to(logits.device, dtype=logits.dtype))
         loss = loss_fct(logits.view(-1, len(LABELS)), labels.view(-1))
 
         return (loss, outputs) if return_outputs else loss
